Fix friends & users links
